Move _import_remote_refs to dulwich.refs.
